requirements.txt
pymeshlab==2025.7.post1
pyvista
pybind11-stubgen
PyOpenGL
PyOpenGL_accelerate
pygame
